requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
torchvision
torchaudio
torch
transformers==4.26.0
diffusers==0.12.0
accelerate
python-multipart
gdown
scikit-image
numpy~=1.24.3
pillow~=9.5.0
tqdm~=4.65.0